added database setup and my sql dependency

+++ src/main/java/org/example/edufyalbumforartist/controllers/ArtistController.java
@@ -18,7 +18,7 @@ public class ArtistController {
     @Autowired
     private ArtistService artistService;
     @GetMapping("/albums/{artistId}")
-    public ResponseEntity<Set<Album>> getAlbumsByArtist(@PathVariable Integer artistId) {
+    public ResponseEntity<Set<Album>> getAlbumsByArtist(@PathVariable Long artistId) {
         return ResponseEntity.ok(artistService.getAlbumsByArtist(artistId));
     }
